refactor: 엔터티 구조 개선 및 공통 책임 분리
- `BasicEntity`를 `BaseUpdatableEntity`와 나누어 수정 가능 속성 분리
- 엔터티 패키지 구조 변경 및 공통 속성 재사용성 향상
- `User`, `ReadStatus`, `UserStatus`, `Channel`, `Message` 클래스가 `BaseUpdatableEntity`를 상속하도록 변경
- 불필요한 복사 생성자 및 관련 메서드 제거
- 테스트 코드에서 사용되지 않는 테스트 케이스 및 주석 제거
